⚙️ FEATURE-#22: Rename IMAPSession to ImapClient; update all imports
