+++ email_profile/advanced.py
@@ -23,7 +23,7 @@ from email_profile.providers import (
 )
 from email_profile.retry import with_retry
 from email_profile.searches import Where
-from email_profile.smtp import SmtpClient
+from email_profile.smtp_client import SmtpClient
 from email_profile.status import IMAPError, Status
 from email_profile.types import SMTPHost
