+++ tests/conftest.py
@@ -81,7 +81,7 @@ def app(fake_client):
     """A connected Email app talking to a fake IMAP server."""
     with (
         patch(
-            "email_profile.imap_session.imaplib.IMAP4_SSL",
+            "email_profile.imap_client.imaplib.IMAP4_SSL",
             return_value=fake_client,
         ),
         Email("imap.x", "u", "p") as connected,
@@ -94,7 +94,7 @@ def gmail_app(gmail_client):
     """A connected Email app with Gmail-style mailbox names."""
     with (
         patch(
-            "email_profile.imap_session.imaplib.IMAP4_SSL",
+            "email_profile.imap_client.imaplib.IMAP4_SSL",
             return_value=gmail_client,
         ),
         Email("imap.x", "u", "p") as connected,
