+++ tests/test_email.py
@@ -9,7 +9,7 @@ class TestConnection(TestCase):
     def test_does_not_connect_eagerly(self):
         fake = make_fake_client()
         with patch(
-            "email_profile.imap_session.imaplib.IMAP4_SSL", return_value=fake
+            "email_profile.imap_client.imaplib.IMAP4_SSL", return_value=fake
         ):
             Email("imap.x", "u", "p")
         fake.login.assert_not_called()
@@ -17,7 +17,7 @@ class TestConnection(TestCase):
     def test_mailbox_requires_connection(self):
         fake = make_fake_client()
         with patch(
-            "email_profile.imap_session.imaplib.IMAP4_SSL", return_value=fake
+            "email_profile.imap_client.imaplib.IMAP4_SSL", return_value=fake
         ):
             app = Email("imap.x", "u", "p")
         with self.assertRaises(NotConnected):
@@ -27,7 +27,7 @@ class TestConnection(TestCase):
         fake = make_fake_client()
         with (
             patch(
-                "email_profile.imap_session.imaplib.IMAP4_SSL",
+                "email_profile.imap_client.imaplib.IMAP4_SSL",
                 return_value=fake,
             ),
             Email("imap.x", "u", "p") as app,
@@ -40,7 +40,7 @@ class TestConnection(TestCase):
         fake = make_fake_client()
         with (
             patch(
-                "email_profile.imap_session.imaplib.IMAP4_SSL",
+                "email_profile.imap_client.imaplib.IMAP4_SSL",
                 return_value=fake,
             ),
             Email("imap.x", "u", "p") as app,
@@ -124,7 +124,7 @@ class TestMailboxProperties(TestCase):
     def setUp(self):
         self.fake = make_fake_client(mailboxes=GMAIL_LIST)
         self._patcher = patch(
-            "email_profile.imap_session.imaplib.IMAP4_SSL",
+            "email_profile.imap_client.imaplib.IMAP4_SSL",
             return_value=self.fake,
         )
         self._patcher.start()
@@ -154,7 +154,7 @@ class TestQueryShortcuts(TestCase):
     def setUp(self):
         self.fake = make_fake_client()
         self._patcher = patch(
-            "email_profile.imap_session.imaplib.IMAP4_SSL",
+            "email_profile.imap_client.imaplib.IMAP4_SSL",
             return_value=self.fake,
         )
         self._patcher.start()
@@ -202,7 +202,7 @@ class TestPublicProperties(TestCase):
         fake = make_fake_client()
         with (
             patch(
-                "email_profile.imap_session.imaplib.IMAP4_SSL",
+                "email_profile.imap_client.imaplib.IMAP4_SSL",
                 return_value=fake,
             ),
             Email("imap.x", "u", "p") as app,
@@ -272,7 +272,7 @@ class TestNoop(TestCase):
         fake = make_fake_client()
         with (
             patch(
-                "email_profile.imap_session.imaplib.IMAP4_SSL",
+                "email_profile.imap_client.imaplib.IMAP4_SSL",
                 return_value=fake,
             ),
             Email("imap.x", "u", "p") as app,
@@ -303,7 +303,7 @@ class TestRestore(TestCase):
             fake = make_fake_client()
             with (
                 patch(
-                    "email_profile.imap_session.imaplib.IMAP4_SSL",
+                    "email_profile.imap_client.imaplib.IMAP4_SSL",
                     return_value=fake,
                 ),
                 Email("imap.x", "u", "p") as app,
@@ -331,7 +331,7 @@ class TestRestore(TestCase):
             fake = make_fake_client()
             with (
                 patch(
-                    "email_profile.imap_session.imaplib.IMAP4_SSL",
+                    "email_profile.imap_client.imaplib.IMAP4_SSL",
                     return_value=fake,
                 ),
                 Email("imap.x", "u", "p") as app,
@@ -357,7 +357,7 @@ class TestRestore(TestCase):
             fake = make_fake_client()
             with (
                 patch(
-                    "email_profile.imap_session.imaplib.IMAP4_SSL",
+                    "email_profile.imap_client.imaplib.IMAP4_SSL",
                     return_value=fake,
                 ),
                 Email("imap.x", "u", "p") as app,
